+++ src/main/storage.go
@@ -1,7 +1,7 @@
 package main
 
 type Storage interface {
-	Shorten(url string, exp int64) (string, error)
+	Shorten(url string, exp uint32) (string, error)
 	ShortURLInfo(eid string) (interface{}, error)
 	UnShorten(eid string) (string, error)
 }
